refactor: restructure MessageManager actions into separate components
- Deleted the old Actions.vue file and replaced it with a new directory structure for action components.
- Introduced individual action components: QuickMessageAction, AiTextImprovementAction, EmojiAction, AudioAction, AttachAction, InternalNoteAction, and SendAction.
- Updated the main TextBox index.vue to reflect the new component structure.
- Adjusted tests to import from the new Actions directory.
