+++ src/components/chats/MessageManagerV2/TextBox/Actions/AudioAction.vue
@@ -1,0 +1,68 @@
+<template>
+  <ActionItem
+    icon="mic"
+    :tooltip="t('audio_message')"
+    :pressed="isPressed"
+    :disabled="isDisabled"
+    :disableFromParent="isAiLoading"
+    @click="handleClick"
+  />
+</template>
+
+<script setup lang="ts">
+import { computed } from 'vue';
+import { storeToRefs } from 'pinia';
+
+import { useMessageManager } from '@/store/modules/chats/messageManager';
+import { useAiTextImprovement } from '@/store/modules/chats/aiTextImprovement';
+
+import ActionItem from './ActionItem.vue';
+
+import i18n from '@/plugins/i18n';
+
+const { t } = i18n.global;
+
+defineOptions({
+  name: 'AudioAction',
+});
+
+const emit = defineEmits<{
+  startAudioRecording: [void];
+}>();
+
+const messageManager = useMessageManager();
+const { clearInputs } = messageManager;
+const {
+  inputMessage,
+  isInternalNote,
+  isSuggestionBoxOpen,
+  audioRecorderStatus,
+  mediaUploadFiles,
+} = storeToRefs(messageManager);
+
+const { isLoading: isAiLoading } = storeToRefs(useAiTextImprovement());
+
+const isValidInputMessage = computed(() =>
+  isSuggestionBoxOpen.value
+    ? !inputMessage.value.startsWith('/')
+    : !!inputMessage.value.trim(),
+);
+
+const isDisabled = computed(
+  () =>
+    isValidInputMessage.value ||
+    mediaUploadFiles.value.length > 0 ||
+    isInternalNote.value,
+);
+
+const isPressed = computed(() =>
+  ['recording', 'recorded', 'playing', 'paused'].includes(
+    audioRecorderStatus.value,
+  ),
+);
+
+function handleClick() {
+  clearInputs();
+  emit('startAudioRecording');
+}
+</script>
