+++ src/components/chats/MessageManagerV2/TextBox/__tests__/TextBox.spec.js
@@ -37,7 +37,7 @@ vi.mock('../AudioRecorder.vue', () => ({
   },
 }));
 
-vi.mock('../Actions.vue', () => ({
+vi.mock('../Actions/index.vue', () => ({
   default: {
     name: 'MessageManagerTextBoxActions',
     template: `<div data-testid="actions">
